Add custom checks for watchman to check cron status Clean up a few imports of local modules Fixes #246
